Setup webpack sw generation + workbox as dependecies :rocket:
