+++ package.json
@@ -61,11 +61,13 @@
     "typescript": "^3.9.2",
     "webpack": "^4.43.0",
     "webpack-fix-style-only-entries": "^0.4.0",
-    "webpack-stream": "^5.2.1"
+    "webpack-stream": "^5.2.1",
+    "workbox-webpack-plugin": "^5.1.3"
   },
   "dependencies": {
     "bootstrap": "^4.5.0",
     "intersection-observer": "^0.10.0",
-    "three": "^0.116.1"
+    "three": "^0.116.1",
+    "workbox-sw": "^5.1.3"
   }
 }
